strip down boilerplate stories
